% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATEk.R
\name{atekCl}
\alias{atekCl}
\title{Estimation of the cluster-specific treatment effects in the partially nested design}
\usage{
atekCl(
  data_in,
  ttname = "tt",
  Kname = "K",
  Yname = "Y",
  Xnames,
  Yfamily = "gaussian",
  learners_tt = c("SL.nnet", "SL.ranger"),
  learners_k = c("SL.nnet.modified"),
  learners_y = c("SL.nnet", "SL.ranger"),
  sensitivity = NULL,
  Fit = "mlr",
  cv_folds = 4L
)
}
\arguments{
\item{data_in}{A \code{data.frame} containing the observed data.
In "data_in", column "tt" is the treatment assignment ("tt" is coded as 0 for individuals in the control arm and as 1 for individuals in the treatment arm);
column "K" is the cluster assignment in the treatment arm ("K" is coded as 1, 2, ..., J for each individual in the treatment arm with J being the number of clusters, and "K" is coded as 0 for individuals in the control arm);}

\item{Xnames}{A character vector of the names of the columns in "data_in" that correspond to baseline covariates (X)}

\item{Ynames}{A character string of the name of the column in "data_in" that correspond to the outcome variable}
}
\description{
Estimation of the cluster-specific treatment effects in the partially nested design
}
\examples{

# library(tidyverse)
# library(SuperLearner)
# library(glue)
#
# data(data_in)
# data_in <- data_in
# Xnames <- c(grep("X_dat", colnames(data_in), value = TRUE))
#
# estimates_ate_K <- atekCl(
# data_in = data_in, ttname = "tt", Kname = "K", Yname = "Y",
# Xnames = Xnames,
# Yfamily = "gaussian",
# sensitivity = NULL # or "small_to_medium"
# )
# estimates_ate_K$ate_K


}
