% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in.R
\docType{data}
\name{data_in}
\alias{data_in}
\title{data_in}
\format{
A data frame with 400 rows and 8 variables:
\describe{
  \item{Y}{Outcome.}
  \item{K}{Cluster assignment in the treatment arm.}
  \item{tt}{Treatment assignment. 1 for individuals assigned to the treatment arm. 0 for individuals assigned to the control arm. The control arm is unclustered.}
  \item{X_dat.1}{Baseline covariates.}
  \item{X_dat.2}{Baseline covariates.}
  \item{X_dat.3}{Baseline covariates.}
  \item{X_dat.4}{Baseline covariates.}
  \item{id}{Individual id.}
}
}
\usage{
data_in
}
\description{
data_in_with_partially_nested_design
}
\keyword{datasets}
